[VOID] Toolkit: Restructured Tools for viewer and annotations into Toolkit
